+++ src/main/DesktopPane.java
@@ -40,6 +40,7 @@ import java.awt.GraphicsDevice;
 import java.awt.GraphicsEnvironment;
 import java.awt.Image;
 import java.awt.Point;
+import java.awt.Desktop;
 import java.awt.Window;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -50,6 +51,7 @@ import java.awt.event.ItemListener;
 import java.awt.event.KeyEvent;
 import java.io.PrintStream;
 import java.lang.reflect.Method;
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -83,6 +85,10 @@ public class DesktopPane extends JFrame implements ActionListener {
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
+
+	public final String projectWebsite = "https://github.com/move-ucsb/DynamoVis";
+	private static final String buildVersionString = "1.0";
+	private static final String buildVersionDate = "Aug 1, 2021";
 
 	JFrame desktop;
 	public Dimension animationSize = new Dimension(1280, 720);
@@ -192,7 +198,10 @@ public class DesktopPane extends JFrame implements ActionListener {
 		menuBar = createMenuBar();
 		setJMenuBar(menuBar);
 
-		setTitle("DynamoVis - Configure Animation");
+		if(isMacOSX())
+			setTitle("Input Data and Animation Configuration");
+		else 
+			setTitle("Input Data and Animation Configuration - [DynamoVis]");
 		setResizable(true);
 		setSize(250, 700);
 		setLocation(sWidth / 5, sHeight / 5);
@@ -204,6 +213,7 @@ public class DesktopPane extends JFrame implements ActionListener {
 			// macos app icon
 			Image icon = new ImageIcon(this.getClass().getClassLoader().getResource("logo1024.png")).getImage();
 			com.apple.eawt.Application.getApplication().setDockIconImage(icon);
+			
 			// System.out.println("Platform: macOS");
 		} else {
 			// windows app icon
@@ -261,7 +271,7 @@ public class DesktopPane extends JFrame implements ActionListener {
 		System.out.println("# DynamoVis Animation Tool");
 		System.out.println("# Copyright (C) 2016 Glenn Xavier");
 		System.out.println("#      Updated: 2021 Mert Toka");
-		System.out.println("# Build 0.5.0-dev, Apr 30, 2021");
+		System.out.println("# Version "+buildVersionString+", "+buildVersionDate+"");
 		System.out.println("# This program comes with ABSOLUTELY NO WARRANTY");
 		System.out.println("# This is free software, and you are welcome to \nredistribute it under certain conditions.");
 		System.out.println("");
@@ -330,6 +340,14 @@ public class DesktopPane extends JFrame implements ActionListener {
 		// Recorder recorder = new Recorder(this,bContainer);//kate added passing in
 		// bContainer
 		recordContainer.setContentPane(recorder);
+		recordContainer.addComponentListener(new ComponentAdapter() {
+			public void componentHidden(ComponentEvent e) {
+				recorder.discardRecorder();
+			}
+			public void componentShown(ComponentEvent e) {
+				recorder.enableRecorder();
+			}
+		});
 		recordContainer.pack();
 
 		if (startup) {
@@ -355,7 +373,7 @@ public class DesktopPane extends JFrame implements ActionListener {
 			pack();
 		}
 
-		setTitle("DynamoVis Animation Tool");
+		// setTitle("DynamoVis Animation Tool");
 
 		dataFilePath = path;
 
@@ -658,15 +676,15 @@ public class DesktopPane extends JFrame implements ActionListener {
 		record.addActionListener(this);
 		export.add(record);
 
-		// JMenu help = new JMenu("Help");
-		// help.setMnemonic(KeyEvent.VK_H);
-		// menuBar.add(help);
+		JMenu help = new JMenu("Help");
+		help.setMnemonic(KeyEvent.VK_H);
+		menuBar.add(help);
 
-		// JMenuItem about = new JMenuItem("About");
-		// about.setActionCommand("about");
-		// about.addActionListener(this);
-		// about.setEnabled(false);
-		// help.add(about);
+		JMenuItem about = new JMenuItem("About");
+		about.setActionCommand("about");
+		about.addActionListener(this);
+		about.setEnabled(true);
+		help.add(about);
 
 		// JMenu dev = new JMenu("Dev");
 		// menuBar.add(dev);
@@ -721,39 +739,62 @@ public class DesktopPane extends JFrame implements ActionListener {
 		return System.getProperty("os.name").indexOf("Mac OS X") >= 0;
 	}
 
+	public static boolean openWebpage(String urlString) {
+		Desktop desktop = Desktop.isDesktopSupported() ? Desktop.getDesktop() : null;
+		if (desktop != null && desktop.isSupported(Desktop.Action.BROWSE)) {
+			try {
+				desktop.browse(new URI(urlString));
+				return true;
+			}
+			catch (Exception e) {
+				e.printStackTrace();
+			}
+		}
+		return false;
+	}
+
 	// ACTION LISTENER OVERRIDE ----------------------
 	@Override
 	public void actionPerformed(ActionEvent e) {
 		switch (e.getActionCommand()) {
-		case "new" -> {
-			if (sketch != null) {
-				sketch.isExiting = true;
-				// sketch.exit();
+			case "new" -> {
+				if (sketch != null) {
+					sketch.isExiting = true;
+					// sketch.exit();
+				}
 			}
-		}
-		case "edit" -> {
-		}
-		case "color" -> {
-		}
-		case "legend" -> {
-		}
-		case "quit" -> quit();
-		case "record" -> recordContainer.setVisible(true);
-		case "basemap" -> baseMapContainer.setVisible(true);
-		case "reset" -> resetWindowLocs();
-		
-		}
-		;
+			case "edit" -> {}
+			case "color" -> {}
+			case "legend" -> {}
+			case "quit" -> quit();
+			case "record" -> recordContainer.setVisible(true);
+			case "basemap" -> baseMapContainer.setVisible(true);
+			case "reset" -> resetWindowLocs();
+			case "about" -> openWebpage(projectWebsite);
+			case "histo" -> {
+				// if (hc == null) {
+				// 	hc = new JDialog(this);
+				// 	hc.setTitle("HISTO TEST");
+				// 	Histo histo = new Histo();
+				// 	hc.setContentPane(histo);
+				// 	hc.setLocationRelativeTo(this);
+				// 	hc.pack();
+				// }
+				// hc.setVisible(true);
+			}
+		};
 	}
 
 	// MAIN ------------------------------
 	public static void main(String[] args) {
+		if (isMacOSX()) {
+			System.setProperty("apple.laf.useScreenMenuBar", "true");
+			System.setProperty("com.apple.mrj.application.apple.menu.about.name", "DynamoVis");
+		}
+		// System.out.println("App started, no UI yet");
 		javax.swing.SwingUtilities.invokeLater(new Runnable() {
 			public void run() {
-				if (isMacOSX()) {
-					System.setProperty("apple.laf.useScreenMenuBar", "true");
-					System.setProperty("com.apple.mrj.application.apple.menu.about.name", "Animation Tool");
-				}
+				// System.out.println("Started the UI");
 				try {
 					UIManager.setLookAndFeel(UIManager.getSystemLookAndFeelClassName());
 				} catch (Exception e) {